check hfs snapshot before destroying it (#796)
* check hfs snapshot before destroying it

* fix formatting

* refactor based on feedback

+++ vps4-orchestration-plugin/src/main/java/com/godaddy/vps4/orchestration/hfs/snapshot/DestroySnapshot.java
@@ -6,6 +6,7 @@ import com.godaddy.vps4.orchestration.snapshot.WaitForSnapshotAction;
 
 import gdg.hfs.orchestration.Command;
 import gdg.hfs.orchestration.CommandContext;
+import gdg.hfs.vhfs.snapshot.Snapshot;
 import gdg.hfs.vhfs.snapshot.SnapshotAction;
 import gdg.hfs.vhfs.snapshot.SnapshotService;
 
@@ -20,14 +21,20 @@ public class DestroySnapshot implements Command<Long, Void> {
 
     @Override
     public Void execute(CommandContext context, Long snapshotId) {
-
-        SnapshotAction hfsAction = context.execute("DestroySnapshotHfs",
-                ctx -> hfsSnapshotService.destroySnapshot(snapshotId),
-                SnapshotAction.class);
-
-        context.execute(WaitForSnapshotAction.class, hfsAction);
-
+        if (doesHfsSnapShotExist(snapshotId, context)) {
+            SnapshotAction hfsAction = context.execute("DestroySnapshotHfs",
+                                                       ctx -> hfsSnapshotService.destroySnapshot(snapshotId),
+                                                       SnapshotAction.class);
+            context.execute(WaitForSnapshotAction.class, hfsAction);
+        }
         return null;
+    }
+
+    private boolean doesHfsSnapShotExist(Long snapshotId, CommandContext context) {
+        Snapshot hfsSnapshot = context.execute("GetSnapshotHfs",
+                                               ctx -> hfsSnapshotService.getSnapshot(snapshotId),
+                                               Snapshot.class);
+        return hfsSnapshot.destroyDate == null && hfsSnapshot.completeDate != null;
     }
 
 }
